extract functions into libraries for contract size reduction

+++ packages/protocol/test/CommentManagerMisc.t.sol
@@ -288,7 +288,7 @@ contract CommentsTest is Test, IERC721Receiver {
       app,
       0,
       commentData.parentId,
-      uint96(block.timestamp),
+      uint88(block.timestamp),
       commentData.content,
       commentData.targetUri,
       commentData.commentType,
